fix: :bug: replace the type property of the objects on array position 0 and 1
The type property should be an array of strings and we have used strings instead. Although this might not seem as an issue, it's actually a big one when it comes to later trying to access those properties. Strings and Arrays though might share some methods, they also have different methods to access them.